:recycle: Refactor common shell env.
This commit makes some changes to the `.zshrc`, `.bashrc` and the custom
dotfile containing common paths, aliases and environment variables for
both shells, `.common`. The file has been split into three different
files, and the changes have been made accordingly. The activation of the
cargo environment was redundant and was removed.