:sparkles: Add + modify scripts.
This commit adds a couple of much needed scripts, and modifies the main
`tochky` script that I use for system administration. The scripts added
are `vol.sh` and `dim.sh` for constructing volume and brightness
indicators respectively (via `dunst`), and `pow.sh` that uses `rofi` to
construct a much simpler powermenu.