:fire: Removing rofi-based applets.
This commit removes all the `rofi`-based applets once and for all. The
only applet that has been kept is the powermenu, since that is the only
things that's useful. It's location was changed in the last commit, and
it was refactored to become much simpler as well. If I add more applets
based on `rofi` in the future, they will go in the `.bin` folder.